Add files via upload
The user will provide the input through web page in rails. Web page should contain a text field (Root directory) and a search button in word files.
Search results will be display on web page as well.
In this Submission for every SSN found, the filename including the full path will be printed to the browser and the content of the document will be printed into a text. Please note the following picture is just an example giving you hints how to achieve the interface. Your interface should show the path and content of word files.